requirements.txt
streamlit
pandas
numpy<2.0.0
scikit-learn
xgboost
Pillow
shap
fasttext-official
joblib